# Minor optimization for building document trees