--- a/tests/Giraffe.ViewEngine.Benchmarks/Giraffe.ViewEngine.Benchmarks.fsproj
+++ b/tests/Giraffe.ViewEngine.Benchmarks/Giraffe.ViewEngine.Benchmarks.fsproj
@@ -1,6 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
+    <OutputType>Exe</OutputType>
   </PropertyGroup>
 
   <ItemGroup>