Changes:
1. Added a benchmark for building the document tree.
2. Removed a redundant ToArray for attributes in XmlElement by changing the collection type from array to list. The public functions (e.g. tag, voidTag, div etc) expect a list anyways.
3. Made XmlElement a struct.

Not sure about the breaking changes policy for this package. Changing the type of XmlElement is technically a breaking change. However, if the user uses the functions provided by ViewEngine - the build will not break. The functions do not expose the details of how XmlElement is constructed.

Results:
Perf: -27% (0.594 us) execution time on building the test document
Memory:  -26% (1.26KB) memory allocations when building the test document

No changes to serialisation benchmark (HtmlUtf8Benchmark)

Benchmark on master (1):
|  Method |     Mean |     Error |    StdDev |   Median | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|----------:|----------:|---------:|------:|-------:|------:|------:|----------:|
| Default | 2.175 us | 0.0434 us | 0.1096 us | 2.132 us |  1.00 | 1.5602 |     - |     - |   4.78 KB |

Benchmark after removing ToArray (2):
|  Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Default | 1.682 us | 0.0285 us | 0.0484 us |  1.00 | 1.3313 |     - |     - |   4.08 KB |

Benchmark after making XmlElement a struct (3):
|  Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Default | 1.581 us | 0.0096 us | 0.0090 us |  1.00 | 1.1463 |     - |     - |   3.52 KB |